Rework AbstractBuilder to remove preview() and add postUpdate().
The preview() method is flawed because in practical use (for evaluating
the default value of a related property) it cannot be guaranteed that
preview() will be called before take(). If it is called afterwards, the
value is not guaranteed to be the same one used by take().

The solution is to provide a reset() method on ValueContainer
that can be called from a postUpdate phase on AbstractBuilder. The value
returned by ValueContainer.get() will remain consistent until reset() is
called.